fix manual tls
